requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
firebase-admin==6.2.0
requests==2.31.0
python-decouple==3.8
spotipy==2.23.0
lyricsgenius==3.0.1
gunicorn==21.2.0
whitenoise==6.6.0

